requirements.txt
fastapi
uvicorn
openai==0.28
tiktoken
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
